Refactored ElevatedView to remove dependency on "Border" and unify all platform under same template
