--- src/Uno.UI.Toolkit/ElevatedView.Android.cs
@@ -1,15 +1,0 @@
-﻿using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-
-namespace Uno.UI.Toolkit
-{
-	partial class ElevatedView : Border
-	{
-		private static void OnChanged(DependencyObject snd, DependencyPropertyChangedEventArgs evt) => ((ElevatedView)snd).UpdateElevation();
-
-		private void UpdateElevation()
-		{
-			this.SetElevationInternal(Background == null ? 0 : Elevation, ShadowColor);
-		}
-	}
-}
